Crate rzif

A z-machine interpreter

The entry point is the main function. Take a look at the binary part of this crate to see an example of how to use. This interpreter follow The Z-Machine Standards Document version 1.1.

Modules

doc
Various documentation

Structs

Config
The configuration of your interface

Error
Information about the error that occured

Enums

Cause
The cause of the error that occured

ErrorLevel
When to report errors

SaveError
Why the restore failed

Trace
A frame of the backtrace

Traits

Interface
Callbacks to your interface

Functions

main
Starts the z-machine interpreter