Enum rzif::Cause

pub enum Cause {
    Quit,
    BadVer,
    TooShort,
    StaticOut,
    WriteOut,
    ReadOut,
    DivByZero,
    MissingOperand,
    BadOpcode,
    StackUnderflow,
    NoVar,
    NoLocal,
    NoLocalInMain,
    MainReturned,
    PcOut,
    TooManyLocals,
    TextBufferOut,
    TerminatingOut,
    StrOut,
    AlphabetOut,
    UnicodeOut,
    BadUnicodeChar,
    BadZSCIIChar,
    NestedAbbr,
    AbbrIncompleteZSCII,
    BadObj,
    NotChildOfParent,
    BadAttr,
    BadProp,
    GetLongProp,
    PutLongProp,
    NoProp,
    BadInputStream,
    BadOutputStream,
    NoOutputS3,
    OutputS3Overflow,
    BadColor,
    SaveInterrupt,
}

The cause of the error that occured

Variants

Quit

The game has quit

BadVer

Storyfile was not in a supported version
data: the version of this storyfile

TooShort

Storyfile too short (it must be at least $40 bytes long to hold the header)
data: length of the storyfile

StaticOut

End of static memory out of bounds
data: end of the static memory

WriteOut

Tried to write outside of dynamic memory
data: address writen

ReadOut

Tried to read out of bounds
data: address readed

DivByZero

Tried to divide by zero

MissingOperand

Not enough operands given data: minimum nuber of operands, current number of operands

BadOpcode

This opcode is invalid
data: operand count, opcode number

StackUnderflow

The stack underflowed

NoVar

Invalid variable id
data: variable id

NoLocal

Non-existent local variable
data: variable id

NoLocalInMain

Access to a local variable from main routine

MainReturned

Main routine returned

PcOut

Program counter out of bounds

TooManyLocals

Routine called that have more than 15 local variables
data: number of local variables

TextBufferOut

Text buffer out of bounds
data: address of the text buffer

TerminatingOut

Terminating charachters table out of bounds
data: address of the table

StrOut

Part of string out of bounds
The address of the string will be in the backtrace

AlphabetOut

Alphabet table out of bounds
data: address of the table

UnicodeOut

Unicode translation table out of bounds
data: address of the table

BadUnicodeChar

Invalid unicode char was encontered
data: unicode char

BadZSCIIChar

Invalid ZSCII char was encontered
data: ZSCII char

NestedAbbr

Abbreviation containing another one
The addresses of the abbreviations will be in the backtrace

AbbrIncompleteZSCII

Abbreviation ending with an incomplete char
The address of the abbreviation will be in the backtrace

BadObj

Invalid object
data: object number

NotChildOfParent

Object not child of his parent
data: number of the child and his parent

BadAttr

Invalid attribute
data: attribute number

BadProp

Invalid property
data: property number

GetLongProp

Got property larger than two bytes
data: object, property

PutLongProp

Set property larger than two bytes
data: object, property

NoProp

Use of a non-existent property
data: object, property

BadInputStream

Invalid output stream
data: stream number

BadOutputStream

Invalid input stream
data: stream number

NoOutputS3

Output stream 3 disabled while not enabled

OutputS3Overflow

output stream 3 enabled for the 17th

BadColor

Invalid color
data: color

SaveInterrupt

Save during interupt\