Function rzif::main

pub fn main<I: Interface>(file: Vec<u8>, config: Config, interface: I)

Starts the z-machine interpreter

The first argument is the contents of the storyfile to play. The second is the configuration of your inferface. The third is the callback functions this crate use to interact with your interface. Before calling this function, the screen must be prepared as described here.
This function returns at the end of the game