Trait rzif::Interface

pub trait Interface {
    fn write_screen(&mut self, text: &str, fixed: bool);

    fn write_transcript(&mut self, text: &str);

    fn write_command(&mut self, text: &str);

    fn status(&mut self, text: &str);

    fn window_font(&mut self, font: u16) -> bool;

    fn window_color(&mut self, foreground: u16, background: u16);

    fn window_style(&mut self, style: u16);

    fn window_set(&mut self, window: u16);

    fn window_buffer(&mut self, buffer: u16);

    fn window_split(&mut self, lines: u16);

    fn window_cursor_set(&mut self, x: u16, y: u16);

    fn window_cursor_get(&mut self) -> (u16, u16);

    fn window_erase(&mut self, window: u16);

    fn window_line(&mut self);

    fn read<F: FnMut(&mut Self) -> bool>(
        &mut self, 
        terminating: Vec<char>, 
        preload: String, 
        maxlen: u16, 
        time: u16, 
        routine: F
    ) -> (String, char);

    fn read_char<T: FnMut(&mut Self) -> bool>(
        &mut self, 
        time: u16, 
        callback: T
    ) -> char;

    fn read_file(&mut self) -> String;

    fn bleep(&mut self, bleep: u16);

    fn save(&mut self, data: &[u8]) -> bool;

    fn restore(&mut self) -> Vec<u8>;

    fn restore_failed(&mut self, cause: SaveError);

    fn error(&mut self, error: Error);
}

Callbacks to your interface

These functions are called when the interpreter needs to do outputs or get inputs.

Required methods

fn write_screen(&mut self, text: &str, fixed: bool)

Print a text to the screen, possibly in fixed-pitch font.

fn write_transcript(&mut self, text: &str)

Write a text to the transcript.

fn write_command(&mut self, text: &str)

Write a text to the commands file.

fn status(&mut self, text: &str)

Write the text in the status line (the top line of the screen).

fn window_font(&mut self, font: u16) -> bool

Set the font. Returns true if the font is available, false otherwise.

fn window_color(&mut self, foreground: u16, background: u16)

Set the current foreground and background true colors.

fn window_style(&mut self, style: u16)

Set the style to apply to the text.

fn window_set(&mut self, window: u16)

Select the current window.

fn window_buffer(&mut self, buffer: u16)

Set the buffer mode.

fn window_split(&mut self, lines: u16)

Split the screen.

fn window_cursor_set(&mut self, x: u16, y: u16)

Move the cursor to (x; y).

fn window_cursor_get(&mut self) -> (u16, u16)

Get the position of the cursor.

fn window_erase(&mut self, window: u16)

Erase a window.

fn window_line(&mut self)

Erase a line.

fn read<F: FnMut(&mut Self) -> bool>(
&mut self,
terminating: Vec<char>,
preload: String,
maxlen: u16,
time: u16,
routine: F
) -> (String, char)

Read a line treminated by a terminating character, with the text preload already given and of maximum length maxlen.

fn read_char<T: FnMut(&mut Self) -> bool>(
&mut self,
time: u16,
callback: T
) -> char

Read a char.

fn read_file(&mut self) -> String

Read a line from a command file.

fn bleep(&mut self, bleep: u16)

Emit the given bleep.

fn save(&mut self, data: &[u8]) -> bool

Save some data into a file. Returns true in case of success, false otherwise.

fn restore(&mut self) -> Vec<u8>

Restore some data from a file. In case of failure an empty Vec can be returned.

fn restore_failed(&mut self, cause: SaveError)

Called if the resore fail

fn error(&mut self, error: Error)

Called on error

Loading content...